Replace OdooRPC with direct requests
Remove the OpenStack Odoo Client for Python's usage of the OdooRPC library, and replace it with direct JSON RPC requests made internally using the `httpx` library.

Functionally, the most important change here is that we now do our own JSON encoding and decoding. This allows us to decode non-integer numbers as `Decimal` objects instead of floats, ensuring that number handling is as accurate as possible (on the client side anyway, Odoo still stores them as floats unfortunately).

There are other benefits, such as performance improvements from avoiding OdooRPC's ORM abstractions and other additional processing we don't need.

+++ openstack_odooclient/managers/credit.py
@@ -16,6 +16,7 @@
 from __future__ import annotations
 
 from datetime import date
+from decimal import Decimal
 from typing import Annotated
 
 from ..base.record import ModelRef, RecordBase
@@ -36,13 +37,13 @@ class Credit(RecordBase["CreditManager"]):
     and caches it for subsequent accesses.
     """
 
-    current_balance: float
+    current_balance: Decimal
     """The current remaining balance on the credit."""
 
     expiry_date: date
     """The date the credit expires."""
 
-    initial_balance: float
+    initial_balance: Decimal
     """The initial balance this credit started off with."""
 
     name: str
